ADD/MOD: Adapted cmake and added samples xml files so they can be merged
CMake will allow options to choose which samples to make and the meta data is
stored in each samples folder. When a sample is enabled the meta data will be
merged with other activated sample's meta data into samples.xml. This file will
be parsed by the samples framework to register demos.